LayoutTests/editing/text-iterator/backwards-text-iterator-attachment-expected.txt

Verify that SimplifiedBackwardsTextIterator treats

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.rangeAsTextUsingBackwardsTextIterator(range) is "some,text"
PASS successfullyParsed is true

TEST COMPLETE
